python-socket/requirements.txt
record3d
Pillow
socket.io
numpy
cv2
